Move all tools from StackRoiWindow to MaskImageWidget (background button, median filter, profile)
TODO: find a way to hide Background button by default
